move _HTTPResponseBody to its own file
