+++ packages/gax/Sources/GoogleCloudGax/HTTPClientResponse.swift
@@ -19,34 +19,6 @@ import struct NIOHTTP1.HTTPHeaders
 import enum NIOHTTP1.HTTPResponseStatus
 import struct NIOCore.ByteBuffer
 
-/// An asynchronous sequence of response body chunks.
-@_spi(GoogleCloudInternal) public struct _HTTPResponseBody: AsyncSequence, Sendable {
-  public typealias Element = Data
-
-  let body: AsyncHTTPClient.HTTPClientResponse.Body
-
-  public init(_ body: AsyncHTTPClient.HTTPClientResponse.Body) {
-    self.body = body
-  }
-
-  public struct AsyncIterator: AsyncIteratorProtocol {
-    var iterator: AsyncHTTPClient.HTTPClientResponse.Body.AsyncIterator
-
-    public init(_ iterator: AsyncHTTPClient.HTTPClientResponse.Body.AsyncIterator) {
-      self.iterator = iterator
-    }
-
-    public mutating func next() async throws -> Data? {
-      guard let buffer = try await self.iterator.next() else { return nil }
-      return Data(buffer: buffer)
-    }
-  }
-
-  public func makeAsyncIterator() -> AsyncIterator {
-    AsyncIterator(self.body.makeAsyncIterator())
-  }
-}
-
 /// Represents an HTTP response.
 ///
 /// The generated code uses this type directly. It exposes the methods we
